+++ crates/github_scbot_cli/src/commands/auth/remove_external_account.rs
@@ -1,6 +1,6 @@
 use argh::FromArgs;
 use async_trait::async_trait;
-use stable_eyre::eyre::Result;
+use github_scbot_sentry::eyre::Result;
 
 use crate::commands::{Command, CommandContext};