+++ crates/github_scbot_cli/src/commands/debug/mod.rs
@@ -2,7 +2,7 @@
 
 use argh::FromArgs;
 use async_trait::async_trait;
-use stable_eyre::eyre::Result;
+use github_scbot_sentry::eyre::Result;
 
 use super::{Command, CommandContext};